# Add `hints.mostly-unused = true` to `aws-sdk-lambda` and `aws-sdk-rds`