A release build of a crate with a dependency on only `aws-sdk-lambda` went from
57s to 40s with this hint.

A release build of a crate with a dependency on only `aws-sdk-rds` went from
1m34s to 49s with this hint.

## Testing

I did not rebuild the binding crates to test this; I tested using the profile
hint to confirm the performance improvement. I *think* I got the correct
binding name for these two bindings, but please confirm.

## Checklist
- [x] For changes to the smithy-rs codegen or runtime crates, I have created a changelog entry Markdown file in the `.changelog` directory, specifying "client," "server," or both in the `applies_to` key.
- [x] For changes to the AWS SDK, generated SDK code, or SDK runtime crates, I have created a changelog entry Markdown file in the `.changelog` directory, specifying "aws-sdk-rust" in the `applies_to` key.

----

_By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice._